+++ TinyNvidiaUpdateChecker/LibaryHandler.cs
@@ -7,6 +7,8 @@ namespace TinyNvidiaUpdateChecker
     class LibaryHandler
     {
 
+        private static bool is64 = Environment.Is64BitOperatingSystem;
+
         public enum Libary
         {
             SEVENZIP,
@@ -46,11 +48,17 @@ namespace TinyNvidiaUpdateChecker
         }
 
         private static LibaryFile Check7Zip() {
+            /* Debug directory */
             if(Directory.Exists("7-Zip")) {
                 LogManager.Log("7-Zip path: " + Path.GetFullPath("7-Zip") + @"\", LogManager.Level.INFO);
                 return new LibaryFile(Path.GetFullPath("7-Zip") + @"\", Libary.SEVENZIP, true);
             }
-            try {
+
+            /* Default installer */
+
+            // amd64 installer on amd64 system, or x86 on x86 system
+            try
+            {
                 using (RegistryKey regKey = Registry.LocalMachine.OpenSubKey(@"SOFTWARE\Microsoft\Windows\CurrentVersion\Uninstall\7-Zip", false)) {
                     LogManager.Log("7-Zip path: " + regKey.GetValue("InstallLocation").ToString(), LogManager.Level.INFO);
                     return new LibaryFile(regKey.GetValue("InstallLocation").ToString(), Libary.SEVENZIP, true);
@@ -58,14 +66,38 @@ namespace TinyNvidiaUpdateChecker
             }
             catch (Exception) { }
 
+            // x86 intaller on amd64 system
+            if (is64) {
+                try {
+                    using (RegistryKey regKey = Registry.LocalMachine.OpenSubKey(@"SOFTWARE\Wow6432Node\Microsoft\Windows\CurrentVersion\Uninstall\7-Zip", false)) {
+                        LogManager.Log("7-Zip path: " + regKey.GetValue("InstallLocation").ToString(), LogManager.Level.INFO);
+                        return new LibaryFile(regKey.GetValue("InstallLocation").ToString(), Libary.SEVENZIP, true);
+                    }
+                }
+                catch (Exception) { }
+            }
+
+            /* MSI installer */
+            
+            // amd64 installer on amd64 system, or x86 on x86 system
             try {
-                using (RegistryKey regKey = Registry.LocalMachine.OpenSubKey(@"SOFTWARE\Wow6432Node\Microsoft\Windows\CurrentVersion\Uninstall\7-Zip", false)) {
-                    LogManager.Log("7-Zip path: " + regKey.GetValue("InstallLocation").ToString(), LogManager.Level.INFO);
-                    return new LibaryFile(regKey.GetValue("InstallLocation").ToString(), Libary.SEVENZIP, true);
+                using (RegistryKey regKey = Registry.LocalMachine.OpenSubKey(@"SOFTWARE\7-Zip", false)) {
+                    LogManager.Log("7-Zip path: " + regKey.GetValue("Path").ToString(), LogManager.Level.INFO);
+                    return new LibaryFile(regKey.GetValue("Path").ToString(), Libary.SEVENZIP, true);
                 }
             }
             catch (Exception) { }
 
+            // x86 intaller on amd64 system
+            if (is64) {
+                try {
+                    using (RegistryKey regKey = Registry.LocalMachine.OpenSubKey(@"SOFTWARE\WOW6432Node\7-Zip", false)) {
+                        LogManager.Log("7-Zip path: " + regKey.GetValue("Path").ToString(), LogManager.Level.INFO);
+                        return new LibaryFile(regKey.GetValue("Path").ToString(), Libary.SEVENZIP, true);
+                    }
+                }
+                catch (Exception) { }
+            }
 
             /* Scoop support */
             string path;
@@ -78,8 +110,26 @@ namespace TinyNvidiaUpdateChecker
                 return new LibaryFile(path, Libary.SEVENZIP, true);
             }
 
-            // installed in program data
+            // installed in Program Data
             path = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.CommonApplicationData), "scoop", "apps", "7zip", "current");
+            if (Directory.Exists(path)) {
+                path += @"\";
+                LogManager.Log("7-Zip path: " + path, LogManager.Level.INFO);
+                return new LibaryFile(path, Libary.SEVENZIP, true);
+            }
+
+            /* Last resort checks */
+
+            // amd64 on amd64 system, or x86 on x86 system
+            path = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.ProgramFiles), "7-Zip");
+            if (Directory.Exists(path)) {
+                path += @"\";
+                LogManager.Log("7-Zip path: " + path, LogManager.Level.INFO);
+                return new LibaryFile(path, Libary.SEVENZIP, true);
+            }
+
+            // x86 on amd64 system
+            path = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.ProgramFilesX86), "7-Zip");
             if (Directory.Exists(path)) {
                 path += @"\";
                 LogManager.Log("7-Zip path: " + path, LogManager.Level.INFO);
@@ -117,7 +167,7 @@ namespace TinyNvidiaUpdateChecker
         }
 
         public override string ToString() {
-            return "InstallLocation: " + InstallLocation + " | libary: " + libary;
+            return "InstallLocation: " + InstallLocation + " | libary: " + libary + " | isInstalled: " + installed;
         }
     }
 }