+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - Implemented a much better dialog for chosing directory
-- Support for alternative locations for 7-Zip (Scoop, x86 install on amd64 systems)
+- Support for alternative locations for 7-Zip (Scoop, x86 install on amd64 systems, MSI installer and last resort)
 - Added check against vendor ID if the gpu drivers aren't installed
 - Added missing text